refactor: merge Actions into Links feature, consolidate controller and reorganize CQRS structure
