+++ Golinks.Application/Features/Links/Commands/RegisterAccess/RegisterAccessCommand.cs
@@ -2,6 +2,6 @@ using Golinks.Application.Common;
 using Golinks.Application.ViewModel;
 using MediatR;
 
-namespace Golinks.Application.Features.Actions.Commands.RegisterAccess;
+namespace Golinks.Application.Features.Links.Commands.RegisterAccess;
 
 public record RegisterAccessCommand(string Slug) : IRequest<Result<LinkViewModel>>;
