+++ Golinks.Application/Features/Links/Queries/GetMetrics/GetMetricsQuery.cs
@@ -1,0 +1,7 @@
+using Golinks.Application.Common;
+using Golinks.Application.ViewModel;
+using MediatR;
+
+namespace Golinks.Application.Features.Links.Queries.GetMetrics;
+
+public record GetMetricsQuery(int PageNumber, int PageSize, int MetricRange, string BaseUrl) : IRequest<Result<PagedResult<LinkMetricViewModel>>>;
